Refactor to seperate files.
